pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
name = "ipybox"
version = "0.0.0"
description = "Python code execution sandbox based on IPython and Docker"
homepage = "https://github.com/gradion-ai/ipybox"
readme = "README.md"
license = "Apache-2.0"
authors = [
    "Martin Krasser <martin@gradion.ai>",
    "Christoph Stumpf <chris@gradion.ai>",
]
packages = [
    { include = "ipybox" },
]

[tool.poetry.dependencies]
aiodocker = "^0.23.0"
aiofiles = "^24.1"
aiohttp = "^3.10"
pillow = "^11.0"
python = "^3.11,<3.14"
tornado = "^6.4"
typer = "^0.12.5"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.48"
mkdocstrings-python = "^1.12.2"

[tool.poetry.group.dev.dependencies]
datamodel-code-generator = "^0.28.5"
fastapi = "^0.115.12"
invoke = "^2.2"
mcp = "^1.6.0"
pre-commit = "^4.0"
pydantic = "^2.11"
pytest = "^8.3"
pytest-asyncio = "^0.24.0"
python-multipart = "^0.0.20"
uvicorn = "^0.34.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.mypy]
python_version = 3.11
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "aiofiles",
    "aiofiles.os",
]
ignore_missing_imports = true

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
pattern = "default-unprefixed"
style = "pep440"
